+++ Changelog.md
@@ -2,10 +2,17 @@
 
 ## Current Version
 
+### v.0.0.20-alpha
+
+* Added ability to use "mkdir" to send "MKD" command to create new directories.
+* Added ability to use "rmdir" to send "RMD" command to delete directories.
+* Added ability to use "rm", "del", or "delete" to send "DELE" command to delete files.
+* Added ability to use "ren", "mv", or "rename" to rename files.
+
+## Previous Versions
+
 ### v.0.0.19-alpha
 
 * Began tracking changes in change log.
 * Modified ftp command entry to accept multiple arguements
 * Added ability to use "cd" command to send "CWD" and change directories.
-
-## Previous Versions